brain/requirements.txt
aiohttp==3.9.3
numpy==1.26.4
pandas==2.2.1
python-dotenv==1.0.1
websockets==12.0
uvloop==0.19.0

